+++ src/main/java/ro/uaic/asli/lab7/repository/GenreRepository.java
@@ -3,5 +3,5 @@ package ro.uaic.asli.lab7.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ro.uaic.asli.lab7.entity.GenreEntity;
 
-public interface GenreRepository extends JpaRepository<GenreEntity, Long> {
+public interface GenreRepository extends JpaRepository<GenreEntity, Integer> {
 }
